+++ .gitignore
@@ -1,2 +1,2 @@
-/.vscode
-/.__pycache__/
+.vscode
+__pycache__/
