+++ config.py
@@ -1,0 +1,2 @@
+class Config(object):
+    SECRET_KEY = "hardPassword"
