--- templates/signup.html
@@ -1,38 +1,0 @@
-{% extends "base.html" %}
-{% block title %}SignUp Page{% endblock %}
-
-{% block content %}
-
-    {% with messages = get_flashed_messages() %}
-        {% if messages %}
-            {% for msgs in messages %}
-                <div class="alert alert-info" role="alert">
-                    {{msgs}}
-                </div>
-            {% endfor %}
-        {% endif %}
-    {% endwith %}
-
-<h1>Create an Account</h1>
-<p><em>All fields are required</em></p>
-
-<form method="POST" style="margin-top: 30px;">
-    <label for="fullname">Full Name</label>
-    <input name="fullname" id="fullname" class="form-control" placeholder="e.g.: Liam William James" required>
-    <p id="passwordHelpBlock" class="form-text text-muted">
-    </p>
-
-    <label for="email">E-mail</label>
-    <input name="email" id="email" type="email" class="form-control" placeholder="your@email.com" required>
-    <p id="passwordHelpBlock" class="form-text text-muted">
-    </p>
-
-    <label for="password">Password</label>
-    <input name="password" id="password" type="password" class="form-control" placeholder="*****************" required>
-    <p id="passwordHelpBlock" class="form-text text-muted">
-    </p>
-
-    <button type="submit" value="submit" class="btn btn-primary btn-lg btn-block" style="margin-top: 40px;">Sign Up</button>
-</form>
-
-{% endblock %}
